Modify the profile image;reset ios BPods
