+++ src/styles/utils/index.ts
@@ -657,17 +657,6 @@ function getButtonStyleWithIcon(
   }
 }
 
-function getSuccessIndicatorStyle(): ViewStyle {
-  return {
-    position: 'absolute',
-    width: variables.successIndicatorSize,
-    height: variables.successIndicatorSize,
-    borderRadius: variables.successIndicatorSize,
-    right: 12,
-    top: variables.contentHeaderHeight + 8,
-  };
-}
-
 /**
  * Returns the font size for the HTML code tag renderer.
  */
@@ -1173,7 +1162,6 @@ const staticStyleUtils = {
   getSafeAreaPadding,
   getSignUpSafeAreaPadding,
   getSignUpLogoWidthStyle,
-  getSuccessIndicatorStyle,
   getTextColorStyle,
   getTransparentColor,
   getWidthAndHeightStyle,
